+++ src/router.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Editor from './views/Editor.vue';
+import Editor from './Editor.vue';
 
 Vue.use(Router);
