+++ src/views/Config/ComponentActions.vue
