+++ src/views/Config/ComponentMessages.vue
