+++ src/views/Config/ConfigComponentWrapper.vue
@@ -37,7 +37,7 @@ import {TConfig} from '@locational/config-validation/build/module/lib/config_typ
 
 import ComponentMessages from './ComponentMessages.vue';
 import ComponentActions from './ComponentActions.vue';
-import {component_list} from '@/views/component_defs';
+import {component_list} from '@/views/Config/component_defs';
 
 interface NodeComponent extends Vue {
   // TODO: Cannot access ConfigNodeMixin for some reason, so recreating the required parts here
