+++ src/views/Config/components/IrsMonitor.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-  import ConfigNodeMixin from '@/lib/ConfigNodeMixin';
+  import ConfigNodeMixin from '../../../lib/ConfigNodeMixin';
 
   export default ConfigNodeMixin.extend({
   });
